fix: change new components to colocated components
Embroider optimized was failing. After looking into the issues, I found
this: https://github.com/embroider-build/embroider/issues/1261

Changing new helper components I created in helpers.ts to
colocated components fixed this.
